refactor(ui): isolate native terminal draft tests
Keep native-start regressions in a focused suite with shared attachment fixtures.
Centralize the repeated current-repository refresh guard in its existing owner.
Preserve startup, destination, and repository-state behavior within file limits.
