+++ ui/src/pages/new-session/draft-submission-flow.test-support.ts
@@ -1,6 +1,7 @@
 import { vi } from "vitest";
 import { createChatSubmissions } from "../../app/chat-submissions.ts";
 import type { ApplicationContext } from "../../app/context.ts";
+import { registerChatAttachmentPayload } from "../chat/attachment-payload-store.ts";
 import { DraftGatewayState } from "./draft-gateway-state.ts";
 import { DraftPlaceBrowser } from "./draft-place-browser.ts";
 import { DraftPlaceState } from "./draft-place-state.ts";
@@ -155,4 +156,20 @@ export function createDraftFixture(options: FixtureOptions = {}) {
   place.setAgentsHydrated(true);
   place.adoptAgentDefaults();
   return { capabilities: flow.capabilities, context, flow, gateway, place, request, requestUpdate };
+}
+
+export function registerTextPayload(id: string) {
+  return registerChatAttachmentPayload({
+    attachment: { id, mimeType: "text/plain", fileName: `${id}.txt` },
+    dataUrl: `data:text/plain;base64,${btoa(id)}`,
+    file: new File([id], `${id}.txt`, { type: "text/plain" }),
+  });
+}
+
+export function stubObjectUrls(...urls: string[]) {
+  const createObjectURL = vi.fn();
+  urls.forEach((url) => createObjectURL.mockReturnValueOnce(url));
+  const revokeObjectURL = vi.fn();
+  vi.stubGlobal("URL", { createObjectURL, revokeObjectURL });
+  return revokeObjectURL;
 }
